+++ drivers/power/supply/Kconfig
@@ -943,13 +943,13 @@ config CHARGER_QCOM_SMB2
 	  adds support for the SMB2 switch mode battery charger found
 	  in PMI8998 and related PMICs.
 
-config BATTERY_PMI8998_FG
-	tristate "Qualcomm PMI8998 fuel gauge driver"
+config BATTERY_QCOM_FG
+	tristate "Qualcomm PMIC fuel gauge driver"
 	depends on MFD_SPMI_PMIC
 	help
-	  Say Y here to enable the Qualcomm PMI8998 Fuel Gauge driver. This 
+	  Say Y here to enable the Qualcomm PMIC Fuel Gauge driver. This 
 	  adds support for battery fuel gauging and state of charge of 
-	  battery connected tothe fuel gauge. The state of charge is 
+	  battery connected to the fuel gauge. The state of charge is 
 	  reported through a BMS power supply property and also sends 
 	  uevents when the capacity is updated.
