+++ drivers/power/supply/Makefile
@@ -110,4 +110,4 @@ obj-$(CONFIG_BATTERY_SURFACE)	+= surface_battery.o
 obj-$(CONFIG_CHARGER_SURFACE)	+= surface_charger.o
 obj-$(CONFIG_BATTERY_UG3105)	+= ug3105_battery.o
 obj-$(CONFIG_CHARGER_QCOM_SMB2)	+= qcom_pmi8998_charger.o
-obj-$(CONFIG_BATTERY_PMI8998_FG)	+= pmi8998_fg.o
+obj-$(CONFIG_BATTERY_QCOM_FG)	+= qcom_fg.o
